+++ buildSrc/src/main/resources/forbidden/all-signatures.txt
@@ -105,3 +105,10 @@ java.lang.Thread#getAllStackTraces()
 java.lang.System#exit(int)
 java.lang.Runtime#exit(int)
 java.lang.Runtime#halt(int)
+
+@defaultMessage Treat system properties as immutable
+java.lang.System#setProperties(java.util.Properties)
+java.lang.System#setProperty(java.lang.String,java.lang.String)
+java.lang.System#clearProperty(java.lang.String)
+java.lang.System#getProperties() @ Use BootstrapInfo.getSystemProperties for a read-only view
+
