+++ core/src/main/java/org/elasticsearch/bootstrap/BootstrapCLIParser.java
@@ -23,6 +23,7 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.elasticsearch.Build;
 import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.SuppressForbidden;
 import org.elasticsearch.common.cli.CliTool;
 import org.elasticsearch.common.cli.CliToolConfig;
 import org.elasticsearch.common.cli.Terminal;
@@ -100,6 +101,8 @@ final class BootstrapCLIParser extends CliTool {
                 .stopAtNonOption(true) // needed to parse the --foo.bar options, so this parser must be lenient
                 .build();
 
+        // TODO: don't use system properties as a way to do this, its horrible...
+        @SuppressForbidden(reason = "Sets system properties passed as CLI parameters")
         public static Command parse(Terminal terminal, CommandLine cli) {
             if (cli.hasOption("V")) {
                 return Version.parse(terminal, cli);
