+++ core/src/main/java/org/elasticsearch/bootstrap/BootstrapInfo.java
@@ -19,9 +19,15 @@
 
 package org.elasticsearch.bootstrap;
 
+import org.elasticsearch.common.SuppressForbidden;
+
+import java.util.Dictionary;
+import java.util.Enumeration;
+
 /** 
  * Exposes system startup information 
  */
+@SuppressForbidden(reason = "exposes read-only view of system properties")
 public final class BootstrapInfo {
 
     /** no instantiation */
@@ -57,4 +63,61 @@ public final class BootstrapInfo {
      * This is not a full URL, just a path.
      */
     public static final String UNTRUSTED_CODEBASE = "/untrusted";
+
+    // create a view of sysprops map that does not allow modifications
+    // this must be done this way (e.g. versus an actual typed map), because
+    // some test methods still change properties, so whitelisted changes must
+    // be reflected in this this view.
+    private static final Dictionary<Object,Object> SYSTEM_PROPERTIES;
+    static {
+        final Dictionary<Object,Object> sysprops = System.getProperties();
+        SYSTEM_PROPERTIES = new Dictionary<Object,Object>() {
+
+            @Override
+            public int size() {
+                return sysprops.size();
+            }
+
+            @Override
+            public boolean isEmpty() {
+                return sysprops.isEmpty();
+            }
+
+            @Override
+            public Enumeration<Object> keys() {
+                return sysprops.keys();
+            }
+
+            @Override
+            public Enumeration<Object> elements() {
+                return sysprops.elements();
+            }
+
+            @Override
+            public Object get(Object key) {
+                return sysprops.get(key);
+            }
+
+            @Override
+            public Object put(Object key, Object value) {
+                throw new UnsupportedOperationException("treat system properties as immutable");
+            }
+
+            @Override
+            public Object remove(Object key) {
+                throw new UnsupportedOperationException("treat system properties as immutable");
+            }
+        };
+    }
+
+    /**
+     * Returns a read-only view of all system properties
+     */
+    public static Dictionary<Object,Object> getSystemProperties() {
+        SecurityManager sm = System.getSecurityManager();
+        if (sm != null) {
+            sm.checkPropertyAccess("*");
+        }
+        return SYSTEM_PROPERTIES;
+    }
 }
