+++ core/src/test/java/org/elasticsearch/benchmark/monitor/os/OsProbeBenchmark.java
@@ -19,10 +19,12 @@
 
 package org.elasticsearch.benchmark.monitor.os;
 
+import org.elasticsearch.common.SuppressForbidden;
 import org.elasticsearch.common.logging.ESLogger;
 import org.elasticsearch.common.logging.ESLoggerFactory;
 import org.elasticsearch.monitor.os.OsProbe;
 
+@SuppressForbidden(reason = "not really source code or a test")
 public class OsProbeBenchmark {
 
     private static final int ITERATIONS = 100_000;
