+++ qa/evil-tests/src/test/resources/org/elasticsearch/tribe/elasticsearch.yml
