+++ test-framework/src/main/java/org/elasticsearch/test/ESTestCase.java
@@ -42,6 +42,7 @@ import org.elasticsearch.bootstrap.BootstrapForTesting;
 import org.elasticsearch.cache.recycler.MockPageCacheRecycler;
 import org.elasticsearch.client.Requests;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
+import org.elasticsearch.common.SuppressForbidden;
 import org.elasticsearch.common.io.PathUtils;
 import org.elasticsearch.common.io.PathUtilsForTesting;
 import org.elasticsearch.common.logging.ESLogger;
@@ -163,6 +164,7 @@ public abstract class ESTestCase extends LuceneTestCase {
     // randomize and override the number of cpus so tests reproduce regardless of real number of cpus
 
     @BeforeClass
+    @SuppressForbidden(reason = "sets the number of cpus during tests")
     public static void setProcessors() {
         int numCpu = TestUtil.nextInt(random(), 1, 4);
         System.setProperty(EsExecutors.DEFAULT_SYSPROP, Integer.toString(numCpu));
@@ -170,6 +172,7 @@ public abstract class ESTestCase extends LuceneTestCase {
     }
 
     @AfterClass
+    @SuppressForbidden(reason = "clears the number of cpus during tests")
     public static void restoreProcessors() {
         System.clearProperty(EsExecutors.DEFAULT_SYSPROP);
     }
